Also apply to queue screen
